Move tasks for adding user saved tracks to accounts tasks
Rename helper class for getting and refreshing SpotifyUserAuth
Move tests for get_and_refresh_spotify_user_auth_record to
SpotifyAuthUserMixin
