+++ apps/accounts/tasks.py
@@ -1,7 +1,10 @@
 from logging import getLogger
 
-from accounts.models import MoodyUser, UserEmotion, UserSongVote
-from base.tasks import MoodyBaseTask
+from celery.schedules import crontab
+from spotify import SpotifyClient, SpotifyException
+
+from accounts.models import MoodyUser, SpotifyUserAuth, UserEmotion, UserSongVote
+from base.tasks import MoodyBaseTask, MoodyPeriodicTask
 from libs.moody_logging import auto_fingerprint, update_logging_data
 from tunes.models import Emotion
 
@@ -104,3 +107,97 @@ class UpdateUserEmotionRecordAttributeTask(MoodyBaseTask):
                 'song_danceability': vote.song.danceability,
             }
         )
+
+
+class SpotifyAuthUserMixin(object):
+
+    def retry(self):
+        """Dummy method for calling celery retry for task. Needed for testing purposes."""
+        return super(SpotifyAuthUserMixin, self).retry()
+
+    @update_logging_data
+    def get_and_refresh_spotify_user_auth_record(self, auth_id, **kwargs):
+        """
+        Fetch the SpotifyUserAuth record for the given primary key, and refresh if
+        the access token is expired
+
+        :param auth_id: (int) Primary key for SpotifyUserAuth record
+
+        :return: (SpotifyUserAuth)
+        """
+        try:
+            auth = SpotifyUserAuth.objects.get(pk=auth_id)
+        except (SpotifyUserAuth.MultipleObjectsReturned, SpotifyUserAuth.DoesNotExist):
+            logger.error(
+                'Failed to fetch SpotifyUserAuth with pk={}'.format(auth_id),
+                extra={'fingerprint': auto_fingerprint('failed_to_fetch_spotify_user_auth', **kwargs)},
+            )
+
+            raise
+
+        if auth.should_update_access_token:
+            try:
+                auth.refresh_access_token()
+            except SpotifyException:
+                logger.warning(
+                    'Failed to update access token for SpotifyUserAuth with pk={}'.format(auth_id),
+                    extra={'fingerprint': auto_fingerprint('failed_to_update_access_token', **kwargs)},
+                )
+                self.retry()
+
+        return auth
+
+
+class CreateSpotifyAuthUserSavedTracksTask(MoodyBaseTask, SpotifyAuthUserMixin):
+    max_retries = 3
+    default_retry_delay = 60 * 15
+
+    @update_logging_data
+    def run(self, user_auth_id, *args, **kwargs):
+        """
+        Fetch the songs the user has saved in their Spotify account and save to their
+        record for use in generating their browse playlist.
+
+        :param user_auth_id: (int) Primary key of SpotifyAuthUser record to fetch
+        """
+        auth = self.get_and_refresh_spotify_user_auth_record(user_auth_id)
+        client = SpotifyClient(identifier='create_spotify_saved_tracks:{}'.format(auth.spotify_user_id))
+
+        try:
+            spotify_track_ids = client.get_user_saved_tracks(auth.access_token)
+            auth.saved_songs = spotify_track_ids
+            auth.save()
+            logger.info(
+                'Successfully updated Spotify saved songs for user {}'.format(auth.spotify_user_id),
+                extra={
+                    'fingerprint': auto_fingerprint('success_get_saved_songs_from_spotify', **kwargs),
+                    'spotify_user_id': auth.spotify_user_id,
+                    'spotify_user_auth_id': auth.pk
+                }
+            )
+        except SpotifyException:
+            logger.warning(
+                'Error fetching user saved songs from Spotify for user {}'.format(auth.spotify_user_id),
+                extra={
+                    'fingerprint': auto_fingerprint('failed_get_saved_songs_from_spotify', **kwargs),
+                    'spotify_user_id': auth.spotify_user_id,
+                    'spotify_user_auth_id': auth.pk
+                }
+            )
+
+            self.retry()
+
+
+class UpdateSpotifyAuthUserSavedTracksTask(MoodyPeriodicTask):
+    run_every = crontab(minute=0, hour=2, day_of_week=0)
+
+    @update_logging_data
+    def run(self, *args, **kwargs):
+        """
+        Periodically update the saved_songs for each of the SpotifyUserAuth records we have stored.
+        This will ensure we keep our concept of user saved tracks on Spotify fresh.
+        """
+        auths = SpotifyUserAuth.objects.all()
+
+        for auth in auths:
+            UpdateSpotifyAuthUserSavedTracksTask().delay(auth.pk)
